# andersen_todo/static/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: "ToDo list"
  description: |-
    ## Common info
    Simple RESTful application for managing a task list (ToDo list) based on Django (DRF)
    and PostreSQL. Made as Python technical task for Andersen.

    There are three options for API testing available:
    * [Swagger-ui](http://localhost/openapi)
    * [Postman collection](http://localhost/static/openapi/andersen_todo.postman_collection.json)<br>
    * run: 'pytest tests.py' (/andersen_todo/tasks/)

    [GitHub](https://github.com/fruit-chili/andersen_todo)

  contact:
    name: "Dzmitry Surmiankou"
    email: "d.surmiankou@gmail.com"
  version: '1.0.0'
servers:
  - url: http://localhost/
    description: local server
...
paths: 
  /registration/:
    post:
      tags:
        - auth
      summary: New user registration
      operationId: registration
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Registration'          
      responses:
        201: 
          description: Successful operation
        400: 
          description: Bad request, see response text   
...
components:
  schemas:
    Registration:
      type: object
      required: [first_name, username, password]
      properties:
        first_name:
          type: string
          description: Users' first name
          maxLength: 150
          example: John
        username:
          type: string
          description: 150 characters or fewer. Letters, digits and @/./+/-/_ only.
          maxLength: 150
          example: JohnDoe
        password:
          type: string
          description: Password (min 6 symbols)
          example: 123qwerty
        last_name:
          type: string
          description: (optional) Users' last name
          maxLength: 150      
          example: Doe
...
  securitySchemes:
    token_auth:
      type: apiKey
      name: Authorization
      in: header
      description: "The value taken from 'access'-value Response (**/login/**) should 
        be prefixed by the string literal 'Bearer', with whitespace separating
         the two strings. Example: <br>
         **<span style='color:red;'>Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0b2tlbl90eXBlIjoiYWNjZXNzIiwiZXhwIjoxNzIxODI3MDAwLCJpYXQiOjE3MjE3NDA2MDAsImp0aSI6IjVkNTRiM2EwODk4ZTRlZmJiM2Y2YTI3OGFhNjIzZWExIiwidXNlcl9pZCI6NTJ9.5Ijp6IH2Ta1uCExj8eWGuMveh632khAJLTk2_QKVhBA</span>**"
security:
- token_auth: []